// src/store/modules/profile/action-types.ts
export enum ActionTypes {
  ONBOARDING_STEP1 = "ONBOARDING_STEP1",
  ONBOARDING_STEP2 = "ONBOARDING_STEP2",
  ONBOARDING_STEP3 = "ONBOARDING_STEP3",
  ONBOARDING_STEP4 = "ONBOARDING_STEP4",
  ONBOARDING_STEP5 = "ONBOARDING_STEP5",
  ONBOARDING_STEP3_DATA = "ONBOARDING_STEP3_DATA",
  CITY_BY_ZIP = "CITY_BY_ZIP",
}